Update build system to CMake

+++ src/dev/cpuio.hpp
@@ -4,7 +4,7 @@
 
 // Thanks to wiki.osdev.org for much of this code.
 
-static inline void outb(uint16_t port, uint16_t data)
+static inline void outb(uint16_t port, uint8_t data)
 {
   asm volatile( "outb %0, %1" : : "a"(data), "Nd"(port));
 }
